+++ Engine-Noisette/Engine-Noisette/InAir.h
@@ -11,6 +11,6 @@ public:
 
 	// Method
 	void Update() override;
-	void TakeDamage(Character*  p_Character, float damage) override;
+	void TakeDamage(Character*  p_Character, int damage) override;
 };
